Better GH2 summarising
